requirements.txt
asgiref==3.8.1
astroid==3.2.2
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dill==0.3.8
distlib==0.3.8
dj-config-url==0.1.1
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
environs==11.0.0
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==25.8.0
filelock==3.15.3
gunicorn==22.0.0
identify==2.5.36
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jmespath==0.10.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
marshmallow==3.21.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pycparser==2.22
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sentry-sdk==2.11.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
whitenoise==6.7.0
